Show errors instead of warnings in a number of places
In some contexts, genuine errors do not stop execution. Namely:
  - Build errors where the user has requested to continue building
  - Project loading errors in IDE mode
  - Contexts where it's difficult to propagate the error
Until now, we have emitted these errors as warnings, using the logic
that something that does not stop execution is not a proper error.
However, it seems more sensible to categorize them as what they really
are, regardless of whether we continue or not.

Task-number: QBS-1818
Change-Id: Id0038e57a55f077e483e0a2eb58ce95e7fad6a86